my-rag-agent-system/services/retrieval-api/requirements.txt
flask>=3.0.0
psycopg2-binary>=2.9.9
pgvector>=0.2.4
openai>=1.0.0
numpy>=1.24.0
